moved meshdata from Collider.h; other QOL changes
