+++ HollowEngine/HollowEngine/src/Hollow/Components/Body.cpp
@@ -28,6 +28,9 @@ namespace Hollow {
 		if (ImGui::TreeNode("Body"))
 		{
 			ImGui::InputFloat("Mass", &mMass);
+			ImGui::InputFloat3("Position", &mPosition[0]);
+			ImGui::InputFloat3("Velocity", &mVelocity[0]);
+			ImGui::InputFloat3("AngularVelocity", &mAngularVelocity[0]);
 			ImGui::TreePop();
 		}
 	}
