+++ HollowEngine/HollowEngine/src/Hollow/Components/Transform.cpp
@@ -20,10 +20,9 @@ namespace Hollow {
 		if (ImGui::TreeNode("Transform"))
 		{
 			ImGui::InputFloat3("Position", &mPosition[0]);
-			// TODO: Find a better way to update transform matrix
-			glm::mat4 mat(1.0f);
-			mat = glm::translate(mat, mPosition);
-			mTransformationMatrix = mat;
+			ImGui::InputFloat3("Scale", &mScale[0]);
+			ImGui::InputFloat3("Rotation", &mRotation[0]);
+		
 			ImGui::TreePop();
 		}
 	}
